+++ icons/es5/Fluorescent.js
@@ -2,15 +2,15 @@ function _extends() { _extends = Object.assign || function (target) { for (var i
 
 import * as React from "react";
 
-function SvgQrCode(props) {
+function SvgFluorescent(props) {
   return /*#__PURE__*/React.createElement("svg", _extends({
     xmlns: "http://www.w3.org/2000/svg",
     width: 24,
     height: 24,
     viewBox: "0 0 24 24"
   }, props), /*#__PURE__*/React.createElement("path", {
-    d: "M3 11h8V3H3v8zm2-6h4v4H5V5zM3 21h8v-8H3v8zm2-6h4v4H5v-4zm8-12v8h8V3h-8zm6 6h-4V5h4v4zm0 10h2v2h-2zm-6-6h2v2h-2zm2 2h2v2h-2zm-2 2h2v2h-2zm2 2h2v2h-2zm2-2h2v2h-2zm0-4h2v2h-2zm2 2h2v2h-2z"
+    d: "M5 9h14v6H5zm6-7h2v3h-2zm6.286 4.399l1.79-1.803 1.42 1.41-1.79 1.802zM11 19h2v3h-2zm6.29-1.29l1.79 1.8 1.42-1.42-1.8-1.79zM3.495 6.01l1.407-1.408L6.69 6.391 5.284 7.798zm-.003 12.066l1.803-1.79 1.409 1.42-1.803 1.79z"
   }));
 }
 
-export default SvgQrCode;
+export default SvgFluorescent;