+++ src/index.js
@@ -185,6 +185,11 @@ export {
   useAbsoluteLayout,
   useFlexLayout,
 } from 'react-table';
+export {
+  default as Skeleton,
+  SkeletonTheme,
+} from 'react-loading-skeleton';
 export { default as Bubble } from './Bubble';
+export { default as Dropzone } from './Dropzone';
 
 export { default as messages } from './i18n';